% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate-bumps.r
\name{ewaff.annotate.bumps}
\alias{ewaff.annotate.bumps}
\title{ewaff.annotate.bumps}
\usage{
ewaff.annotate.bumps(bumps, annotation)
}
\arguments{
\item{annotation}{A vector annotating the features
in the input of \link{ewaff.bumps}.}
}
\value{
A vector providing the annotation for each bump.

methylation <- ... ## methylation matrix
annotation <- ... ## data frame with gene annotations for each gene
bumps <- ewaff.bumps(..., methylation=methylation, ...)
idx <- match(rownames(methylation), annotation$cpg)
bumps$gene <- ewaff.annotate.bumps(bumps, annotation$gene[idx])
}
\description{
Annotate bumps.
}
