% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bumps.r
\name{ewaff.bumps}
\alias{ewaff.bumps}
\title{ewaff.bumps}
\usage{
ewaff.bumps(estimate, se, p.value, methylation, chr, pos, maxgap = 500,
  p.cutoff = 0.05, verbose = T)
}
\arguments{
\item{estimate}{Vector of EWAS effect estimates (corresponds to rows of \code{methylation}).}

\item{se}{Vector of standard errors of the coefficients.}

\item{p.value}{Vector of p-values.}

\item{methylation}{Methylation matrix (rows=features, columns=samples).}

\item{chr}{Feature chromosome (corresponds to rows of \code{methylation}).}

\item{pos}{Feature chromosome position.}

\item{maxgap}{Maximum distance between consecutive features (Default: 500bp).}

\item{p.cutoff}{Unadjusted p-value cutoff for membership in a candidate hump (Default: 0.05).}

\item{verbose}{If \code{TRUE} (default), then output status messages.}
}
\value{
A data frame listing all candidate bumps and their summary statistics.
}
\description{
Identify differentially methylated regions from EWAS summary statistics.
}
\examples{
chr <- ... ## chromosome of each CpG site
pos <- ... ## position of each CpG site
data <- ... ## data frame containing variable of interest and covariates
methylation <- ... ## methylation matrix
ewas.ret <- ewaff.sites(methylation ~ variable + .,
                      variable.of.interest="variable",
                      methylation=methylation,
                      data=data)

bumps <- ewaff.bumps(estimate=ewas.ret$table$estimate, se=ewas.ret$table$se, p.value=ewas.ret$table$p.value,
                     methylation=methylation,
                     chr=chr, pos=pos)
                     
bumps[which(bumps$p.adjust < 0.05),
      c("chr","start","end","n", "z", "p.value","p.adjust")]

}
