% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpg-summary.r
\name{ewaff.cpg.summary}
\alias{ewaff.cpg.summary}
\title{Summarize CpG distributions}
\usage{
ewaff.cpg.summary(methylation, cpgs.are.rows = T)
}
\arguments{
\item{methylation}{DNA methylation matrix, one row per CpG site,
one column per sample.}

\item{cpgs.are.rows}{logical argument to allow CpG sites to be in 
columns instead of rows.}

\item{...}{Arguments to \code{\link{glm}}.}
}
\value{
matrix with columns of each requested summary statistics 
and row for each CpG in \code{methylation}
}
\description{
Summarize CpG site quantiles (0, 10, 15, 50, 75, 90, 100), the 
mean, and SD.
}
