% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{ewaff.glm.plot}
\alias{ewaff.glm.plot}
\title{Plots GLM regression of a CpG site}
\usage{
ewaff.glm.plot(variable.of.interest, data, methylation, title)
}
\arguments{
\item{variable.of.interest}{Name in \code{colnames(data)} for the variable of interest.}

\item{data}{Data frame containing all variables.}

\item{methylation}{Vector of methylation levels.}

\item{title}{Title of the plot.}

\item{\code{\link{ggplot}}}{object showing the scatterplots
(for continuous variables) or boxplots (for categorical variables) of methylation
vs the variable of interest.  Each plot corresponds to a covariate set.
Methylation levels are in fact residuals from fitting a model with DNA methylation
and the covariates.}
}
\description{
Plots GLM regression of a CpG site
}
