% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{ewaff.glm.residuals.plot}
\alias{ewaff.glm.residuals.plot}
\title{Plots GLM regression residuals vs fitted values for a CpG site}
\usage{
ewaff.glm.residuals.plot(variable.of.interest, data, methylation, title)
}
\arguments{
\item{variable.of.interest}{Name in \code{colnames(data)} for the variable of interest.}

\item{data}{Data frame containing all variables.}

\item{methylation}{Vector of methylation levels.}

\item{title}{Title of the plot.}

\item{\code{\link{ggplot}}}{object showing the scatterplot
of fitted values vs residuals for the model \code{methylation ~ variable + covariates}.}
}
\description{
Plots GLM regression residuals vs fitted values for a CpG site
}
\examples{
methylation <- ... ## methylation matrix
data <- ... ## variable of interest (smoking) and covariates
obj <- ewaff.sites(methylation ~ smoking + ., methylation, data)
cpg <- "cg05575921"
p <- ewaff.glm.residuals.plot("smoking", as.data.frame(object$design), methylation[cpg,], cpg)
print(p)

}
