% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-outliers.r
\name{ewaff.handle.outliers}
\alias{ewaff.handle.outliers}
\title{Handle outliers}
\usage{
ewaff.handle.outliers(
  methylation,
  method = c("iqr", "winsorize"),
  iqr.limit = 3,
  winsorize.pct = 0.05
)
}
\arguments{
\item{methylation}{DNA methylation matrix, one row per CpG site,
one column per sample.}

\item{method}{The method for handling outliers.  Options are 'iqr' (default) and 'winsorize'.
The 'iqr' option sets observations beyond a user-defined mutliple of the inter-quartile
(\code{iqr.limit}) to \code{NA}.  The 'winsorize' option sets observations beyond
the \code{winsorize.pct} or \code{1-winsorize.pct} percentiles to the corresponding
percentile.}

\item{iqr.limit}{The multiple of IQR beyond which observations
should be set to NA. Default is 3.}

\item{winsorize.pct}{The percentile used for winsorizing observations. Default is 0.05.}
}
\value{
List containing the methylation matrix with outliers modified
depending on the method selected, and a log of how many observations were
identified for each probe.
}
\description{
Handle outliers
}
