% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{ewaff.manhattan.plot}
\alias{ewaff.manhattan.plot}
\title{Manhattan plot}
\usage{
ewaff.manhattan.plot(chr, pos, estimates, p.values,
  sig.threshold = 1e-07, title = "Manhattan plot")
}
\arguments{
\item{chr}{Chromosome.}

\item{pos}{Position.}

\item{estimates}{A vector of estimates.}

\item{p.values}{A vector of p-values, one per association test.}

\item{sig.threshold}{P-value threshold for significance (Default: 1e-7).}

\item{title}{Title for the plot (Default: "Manhattan plot").}
}
\value{
\code{\link{ggplot}} showing the Manhattan plot.
}
\description{
Manhattan plot
}
