% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-bumps.r
\name{ewaff.meta.bumps}
\alias{ewaff.meta.bumps}
\title{ewaff.meta.bumps}
\usage{
ewaff.meta.bumps(meta.estimate, meta.p.value, estimate, se, methylation,
  chr, pos, maxgap = 500, p.cutoff = 0.05, verbose = T)
}
\arguments{
\item{meta.estimate}{Vector of meta-analysed effect estimates.}

\item{estimate}{Matrix of effect estimates (rows=features, cols=analyses).}

\item{se}{Matrix of standard errors of the coefficients (rows=features, cols=analyses).}

\item{methylation}{Reference methylation matrix (rows=features, columns=samples).}

\item{chr}{Feature chromosome.}

\item{pos}{Feature chromosome position.}

\item{maxgap}{Maximum distance between consecutive features (Default: 500bp).}

\item{p.cutoff}{Unadjusted p-value cutoff for membership in a candidate hump (Default: 0.05).}

\item{verbose}{If \code{TRUE} (default), then output status messages.}

\item{meta.se}{Vector of meta-analysed errors of the coefficients.}
}
\value{
A data frame listing all candidate bumps and their meta-analysed statistics.
}
\description{
Identify differentially methylated regions from meta-analyzed summary statistics.
}
\examples{
estimates <- ...   ## rows=features, columns=analyses
ses <- ...         ## rows=features, columns=analyses
methylation <- ... ## rows=features, columns=samples
chr <- ...         ## feature chromosome
pos <- ...         ## feature position

ma.ret <- ewaff.meta.sites(estimates, ses, method="FE")

bumps <- ewaff.meta.bumps(ma.ret$estimate, ma.ret$se, estimates, ses, methylation, chr, pos)

bumps[which(bumps$p.adjust < 0.05),
      c("chr","start","end","n", "z", "p.value","p.adjust")]
}
