% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-sites.r
\name{ewaff.meta.sites}
\alias{ewaff.meta.sites}
\title{ewaff.meta.sites}
\usage{
ewaff.meta.sites(estimate, se, ..., verbose = T)
}
\arguments{
\item{estimate}{Matrix of regression estimates (rows=features, columns=samples).}

\item{se}{Matrix of standard errors of the estimates (rows=features, columns=samples).}

\item{...}{Arguments to \code{\link{metafor::rma.uni}}.}
}
\value{
Data frame containing meta-analyzed statistics and heterogeneity measures.
}
\description{
Meta-analyse EWAS summary statistics.
}
\details{
Note: \code{\link{mclapply}} is used to fit meta-analysis models
using multiple cores.
}
