% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{ewaff.qq.plot}
\alias{ewaff.qq.plot}
\title{QQ plot}
\usage{
ewaff.qq.plot(
  p.values,
  sig.threshold = 1e-07,
  sig.color = "red",
  title = "QQ plot",
  xlab = bquote(-log[10]("expected p-values")),
  ylab = bquote(-log[10]("observed p-values")),
  lambda.method = "median"
)
}
\arguments{
\item{p.values}{A vector of p-values, one per association test.}

\item{sig.threshold}{P-value threshold for significance (Default: 1e-7).}

\item{sig.color}{Color for points corresponding to significant tests (Default: "red").}

\item{title}{Title for the plot (Default: "QQ plot").}

\item{xlab}{Label for the x-axis (Default: -log_10(expected p-values)).}

\item{ylab}{Label for the y-axis (Default: -log_10(observed p-values)).}

\item{lambda.method}{Method for calculating genomic inflation lambda.
Valid values are "median", "regression", "robust", or "none" (Default: "median"). A value of "none" will prevent the lambda value being displayed on the plot.}
}
\value{
An \code{\link{ggplot}} object.
}
\description{
QQ plot
}
