% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{ewaff.report.parameters}
\alias{ewaff.report.parameters}
\title{Specify parameters for report}
\usage{
ewaff.report.parameters(sig.threshold = NA, max.plots = 10,
  qq.inflation.method = "median")
}
\arguments{
\item{sig.threshold}{P-value threshold for significance (Default: NA).
If NA, then threshold used will be 0.05 divided by the number of tests/probes.}

\item{max.plots}{Maximum number of plots to generate (Default: 10).}

\item{qq.inflation.method}{Method for calculating genomic inflation lambda.
Valid values are "median", "regression" or "robust" (Default: "median").}
}
\value{
List of parameter values
}
\description{
Specify parameters for report
}
