% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{ewaff.report}
\alias{ewaff.report}
\title{Generate CpG associations report}
\usage{
ewaff.report(object, output.file = "report.html", author = "Analyst",
  study = "Illumina methylation data", ...)
}
\arguments{
\item{object}{Output from \code{\link{ewaff.summary}}.}

\item{output.file}{Default = "ewas-report.html".
If the file extension is not .htm, .html, .HTM or .HTML then
output will be in markdown format.}

\item{author}{Default = "Analyst". Author name to be specified on report.}

\item{study}{Default = "Illumina methylation data". Study name to be specified on report.}

\item{...}{Arguments to be passed to \code{\link{knitr::knit}}}
}
\description{
Generate HTML file that summarises CpG site associations.
}
