% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-summaries.r
\name{ewaff.sample.characteristics}
\alias{ewaff.sample.characteristics}
\title{Describe samples using the variable of interest and covariates.}
\usage{
ewaff.sample.characteristics(variable.of.interest, data)
}
\arguments{
\item{variable.of.interest}{Name of the variable of interest in \code{colnames(data)}.}

\item{data}{Data frame of variables.}
}
\value{
A data frame with one row for each continuous or ordinal variable
and one row for each level of each categorical variable.
In the first case, each row provides the name, mean value and standard deviation of each variable.
In the second case (categorical), each row provides the name of the variable level and
the number of cases and percentage of cases at that level.
}
\description{
Describe samples using the variable of interest and covariates.
}
