% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites.r
\name{ewaff.sites}
\alias{ewaff.sites}
\title{Test associations at CpG sites}
\usage{
ewaff.sites(
  formula,
  variable.of.interest,
  methylation,
  data,
  family = gaussian,
  method = "glm",
  generate.confounders = NULL,
  n.confounders = NULL,
  most.variable = NULL,
  random.subset = 0.05,
  ...,
  debug = F
)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}}:
a symbolic description of the model to be fitted.
DNA methylation is referred to by the variable name 'methylation'.}

\item{variable.of.interest}{Name of variable(s) in the model formula
for which to save summary statistics.  If it is the dependent variable
in the formula, then it must be numeric or binary; otherwise, it may
be any type of variable or even a vector of variables.
The value is ignored if \code{method == "coxph"}.}

\item{methylation}{DNA methylation matrix, one row per CpG site,
one column per sample.}

\item{data}{Data frame of variables to include in the model.}

\item{family}{See description for \code{\link{glm}}.}

\item{method}{Method for regressions: "glm", "rlm", "limma" or "coxph" (Default: "glm").}

\item{generate.confounders}{Generate variables from the methylation data
to adjust for unknown confounders. May be \code{NULL} for none (default),
"sva" or "smartsva" to generate surrogate variables, or "pca" to generate
prinicipal components. If \code{method == "coxph"}, then \code{generate.confounders}
must be \code{NULL} or \code{"pca"}.}

\item{n.confounders}{Number of unknown confounders to generate.  A value of
\code{NULL} allows \code{\link{sva}} to select the number automatically
or if principal components the maximum number (Default: NULL).}

\item{most.variable}{Generate confounders from the
#' given most variable CpG sites rather than the whole matrix (Default: NULL).}

\item{random.subset}{Generate surrogate variables from the
given percentage of randomly selected CpG sites rather than the whole matrix (Default: 0.05, i.e. 5 percent).}

\item{...}{Arguments to \code{\link{glm}}, \code{\link{rlm}}, \code{\link{limma::eBayes}},
or \code{\link{survival::coxph}}.}
}
\value{
List containing a table of association statistics ('table') and the
model design matrix ('design').
}
\description{
Fit generalized linear model (GLM) to methylation levels each CpG site.
}
\details{
Note: \code{\link{mclapply}} is used to fit regression models
using multiple cores.
}
