% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{ewaff.summary}
\alias{ewaff.summary}
\title{Summarize results.}
\usage{
ewaff.summary(
  object,
  chr,
  pos,
  methylation,
  selected.cpg.sites = character(0),
  additional.variables = NULL,
  parameters = ewaff.report.parameters(),
  verbose = T
)
}
\arguments{
\item{object}{Object returned by \code{\link{ewaff.sites}}.}

\item{chr}{Chromosome of each CpG site in \code{methylation}.}

\item{pos}{Position of each CpG site.}

\item{methylation}{Matrix of methylation levels used in the analysis.}

\item{selected.cpg.sites}{Vector of CpG site names to plot (Default: character(0)).}

\item{additional.variables}{A data frame of variables to show associations with the variable of interest.}

\item{parameters}{Default = ewaff.report.parameters(). List of parameter values.
See \code{\link{ewaff.report.parameters}()}.}
}
\value{
List
}
\description{
Generates variable and covariate summary tables, QQ plots,
Manhattan plots, a list of associations, plots of the strongest
associations and plots of selected CpG sites.
}
