% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sva.r
\name{ewaff.sva}
\alias{ewaff.sva}
\title{Generate surrogate variables using the ewaff.sites syntax}
\usage{
ewaff.sva(formula, variable.of.interest, methylation, data,
  family = gaussian, n.sv = 10, most.variable = most.variable,
  algorithm = "sva", ...)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}}:
a symbolic description of the model to be fitted.
DNA methylation is referred to by the variable name 'methylation'.}

\item{variable.of.interest}{Name of variable in the GLM model formula
for which to save summary statistics.}

\item{methylation}{DNA methylation matrix, one row per CpG site,
one column per sample.}

\item{family}{(Default: gaussian).}

\item{n.sv}{Number of surrogate variables to compute (Default: 10).}

\item{most.variable}{Generate surrogate variables from the
given most variable CpG sites rather than the whole matrix (Default: NULL).}

\item{algorithm}{Algorithm to use, either "sva" or "smartsva" (Default: "sva").}
}
\value{
Matrix containing surrogate variables computed
}
\description{
Provide a function to manually compute surrogate variables, using  
the same syntax for submission of a ewaff.sites call
}
\details{
Note: surrogate variables are computed using \code{\link{sva}}
}
